refactor(macos): extract playback-session reporting from AppModel into AppModel+NowPlaying.swift
Continues the AppModel god-object split. Relocates the scrobbling, Now-Playing
inspector details (track people + lyrics), end-of-track handling, and the 1 Hz
status poll — three cohesive MARK sections — into a new `extension AppModel`
file as one contiguous block.

Pure relocation: the 353 moved lines are byte-identical to what was removed.
AppModel.swift drops 5,581 -> 5,227 lines.

Eight access promotions, all private -> internal and each genuinely used
cross-file by a moved method:
- methods called from the init / session lifecycle that stay on AppModel:
  handleTrackEnded (AudioEngine onTrackEnded callback), startPolling / stopPolling
  (session login / logout).
- stored props the moved methods read/write (instance storage can't move to an
  extension): scrobbleGate, currentTrackPeopleForId, currentLyricsForId,
  trackAnnounceTask, didApplyMenuBarPlayingState.
driveScrobble stays private (only used within the band); pollTimer /
currentTrackPeople / currentLyrics / autoplayWhenQueueEnds were already internal.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
